added:
Journal
Logback
Interfaces
Options expiration
Maintenance page
Something else...

+++ src/main/java/options/papertrading/dto/option/OptionDto.java
@@ -12,5 +12,5 @@ public class OptionDto {
     private int daysToMaturity;
     private double buyCollateral;
     private double writeCollateral;
-    private byte buyOrWrite;
+    private int buyOrWrite;
 }
